requirements.txt
git+git://github.com/japaniel/insightvm-python.git@main
python-dotenv
pandas
